Add new ProgressDialog component
