+++ src/screens/Settings/components/Tools/index.tsx
@@ -1,6 +1,6 @@
 import './index.css'
 
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 
 import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
@@ -8,8 +8,7 @@ import { getGameInfo, quoteIfNecessary } from 'src/helpers'
 
 import { ipcRenderer } from 'src/helpers'
 import { Runner } from 'src/types'
-import { Dialog, DialogContent, DialogHeader } from 'src/components/UI/Dialog'
-import { LinearProgress } from '@mui/material'
+import { ProgressDialog } from 'src/components/UI/ProgressDialog'
 
 interface Props {
   appName: string
@@ -21,9 +20,6 @@ export default function Tools({ appName, runner }: Props) {
   const [winecfgRunning, setWinecfgRunning] = useState(false)
   const [winetricksRunning, setWinetricksRunning] = useState(false)
   const [progress, setProgress] = useState<string[]>([])
-  const winetricksOutputBottomRef = useRef<HTMLDivElement>(null)
-  const logRef = useRef<HTMLDivElement>(null)
-  const [autoScroll, setAutoScroll] = useState(true)
 
   type Tool = 'winecfg' | 'winetricks' | string
   async function callTools(tool: Tool, exe?: string) {
@@ -60,32 +56,6 @@ export default function Tools({ appName, runner }: Props) {
   useEffect(() => {
     setProgress([])
   }, [winetricksRunning])
-
-  const scrollToBottom = () => {
-    winetricksOutputBottomRef.current?.scrollIntoView({ behavior: 'auto' })
-  }
-
-  useEffect(() => {
-    if (autoScroll) {
-      scrollToBottom()
-    }
-  }, [progress, autoScroll])
-
-  const onLogScroll = (ev: Event) => {
-    const target = ev.target as HTMLDivElement
-
-    const atTheBottom =
-      target.scrollTop + target.getBoundingClientRect().height >=
-      target.scrollHeight
-
-    setAutoScroll(atTheBottom)
-  }
-
-  useEffect(() => {
-    if (logRef.current) {
-      logRef.current.addEventListener('scroll', onLogScroll)
-    }
-  }, [logRef.current])
 
   const handleRunExe = async () => {
     let exe = ''
@@ -129,48 +99,13 @@ export default function Tools({ appName, runner }: Props) {
     <>
       <div data-testid="toolsSettings" className="settingsTools">
         {winetricksRunning && (
-          <Dialog
+          <ProgressDialog
+            title={'Winetricks'}
+            progress={progress}
             onClose={() => {
               return
             }}
-            className={classNames('WinetricksProgress__dialog')}
-          >
-            <DialogHeader
-              onClose={() => {
-                return
-              }}
-            >
-              <div>Winetricks</div>
-            </DialogHeader>
-            <DialogContent>
-              <div>{t('progress', 'Progress')}:</div>
-              <div className="winetricks log-box" ref={logRef}>
-                {progress.map((line, key) => {
-                  if (line.toLowerCase().includes(' err')) {
-                    return (
-                      <p key={key} className="winetricks log-error">
-                        {line}
-                      </p>
-                    )
-                  } else if (line.toLowerCase().includes(' warn')) {
-                    return (
-                      <p key={key} className="winetricks log-warning">
-                        {line}
-                      </p>
-                    )
-                  } else {
-                    return (
-                      <p key={key} className="winetricks log-info">
-                        {line}
-                      </p>
-                    )
-                  }
-                })}
-                <div ref={winetricksOutputBottomRef} />
-              </div>
-              <LinearProgress />
-            </DialogContent>
-          </Dialog>
+          />
         )}
         <div className="toolsWrapper">
           <button
